backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pdfplumber>=0.11.0
pydantic[email]>=2.7.0
httpx>=0.27.0
requests